feat: add details info
