+++ src/components/Details/Details.module.scss
@@ -1,0 +1,11 @@
+.details {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  border: 1px solid #fff;
+  border-radius: 8px;
+  flex: 1 0 50%;
+  padding: 8px;
+  margin-top: 32px;
+}
