+++ src/components/List/Details/Details.module.scss
@@ -1,14 +1,7 @@
 .card {
-  position: relative;
   color: #fff;
   flex: 1 0 30%;
   border: 1px solid #fff;
   padding: 8px;
   border-radius: 8px;
-
-  .link {
-    position: absolute;
-    display: block;
-    inset: 0;
-  }
 }
