+++ src/components/List/Details/Details.tsx
@@ -1,4 +1,3 @@
-import { Link, generatePath } from 'react-router-dom';
 import { Character } from '../../../types';
 import styles from './Card.module.scss';
 
@@ -8,11 +7,8 @@ const localDate = new Intl.DateTimeFormat('en-GB', {
   year: 'numeric',
 });
 
-export function Card({ name, created, mass, skin_color, url }: Character) {
+export function Details({ name, created, mass, skin_color }: Character) {
   const joinedDate = localDate.format(new Date(created));
-
-  const urlObj = new URL(url);
-  const username = urlObj.pathname.split('/')[3];
 
   return (
     <div className={styles.card}>
@@ -20,12 +16,6 @@ export function Card({ name, created, mass, skin_color, url }: Character) {
       <span>{joinedDate}</span>
       <div>mass - {mass}</div>
       <div>skin color - {skin_color}</div>
-      <Link
-        className={styles.link}
-        to={generatePath('/search/:id', {
-          id: username,
-        })}
-      />
     </div>
   );
 }
